refactor: move base entity to common/entity
