+++ src/main/kotlin/com/wafflestudio/csereal/core/about/database/AboutEntity.kt
@@ -1,6 +1,6 @@
 package com.wafflestudio.csereal.core.about.database
 
-import com.wafflestudio.csereal.common.config.BaseTimeEntity
+import com.wafflestudio.csereal.common.entity.BaseTimeEntity
 import com.wafflestudio.csereal.common.controller.AttachmentContentEntityType
 import com.wafflestudio.csereal.common.controller.MainImageContentEntityType
 import com.wafflestudio.csereal.common.enums.LanguageType
