+++ src/main/kotlin/com/wafflestudio/csereal/core/about/database/StatEntity.kt
@@ -1,6 +1,6 @@
 package com.wafflestudio.csereal.core.about.database
 
-import com.wafflestudio.csereal.common.config.BaseTimeEntity
+import com.wafflestudio.csereal.common.entity.BaseTimeEntity
 import com.wafflestudio.csereal.core.about.dto.FutureCareersStatDegreeDto
 import jakarta.persistence.Entity
 import jakarta.persistence.EnumType
