+++ src/main/kotlin/com/wafflestudio/csereal/core/academics/database/ScholarshipLanguageEntity.kt
@@ -1,6 +1,6 @@
 package com.wafflestudio.csereal.core.academics.database
 
-import com.wafflestudio.csereal.common.config.BaseTimeEntity
+import com.wafflestudio.csereal.common.entity.BaseTimeEntity
 import jakarta.persistence.*
 
 @Entity(name = "scholarship_language")
