+++ src/main/kotlin/com/wafflestudio/csereal/core/admissions/database/AdmissionsEntity.kt
@@ -1,6 +1,6 @@
 package com.wafflestudio.csereal.core.admissions.database
 
-import com.wafflestudio.csereal.common.config.BaseTimeEntity
+import com.wafflestudio.csereal.common.entity.BaseTimeEntity
 import com.wafflestudio.csereal.common.enums.LanguageType
 import com.wafflestudio.csereal.common.utils.cleanTextFromHtml
 import com.wafflestudio.csereal.core.admissions.api.req.AdmissionReqBody
