+++ src/main/kotlin/com/wafflestudio/csereal/core/conference/database/ConferencePageEntity.kt
@@ -1,6 +1,6 @@
 package com.wafflestudio.csereal.core.conference.database
 
-import com.wafflestudio.csereal.common.config.BaseTimeEntity
+import com.wafflestudio.csereal.common.entity.BaseTimeEntity
 import com.wafflestudio.csereal.core.user.database.UserEntity
 import jakarta.persistence.*
