+++ src/main/kotlin/com/wafflestudio/csereal/core/member/database/MemberLanguageEntity.kt
@@ -1,6 +1,6 @@
 package com.wafflestudio.csereal.core.member.database
 
-import com.wafflestudio.csereal.common.config.BaseTimeEntity
+import com.wafflestudio.csereal.common.entity.BaseTimeEntity
 import com.wafflestudio.csereal.core.member.type.MemberType
 import jakarta.persistence.*
