+++ src/main/kotlin/com/wafflestudio/csereal/core/member/database/ProfessorEntity.kt
@@ -1,6 +1,6 @@
 package com.wafflestudio.csereal.core.member.database
 
-import com.wafflestudio.csereal.common.config.BaseTimeEntity
+import com.wafflestudio.csereal.common.entity.BaseTimeEntity
 import com.wafflestudio.csereal.common.controller.MainImageContentEntityType
 import com.wafflestudio.csereal.common.enums.LanguageType
 import com.wafflestudio.csereal.common.utils.StringListConverter
