+++ src/main/kotlin/com/wafflestudio/csereal/core/news/database/TagInNewsEntity.kt
@@ -1,6 +1,6 @@
 package com.wafflestudio.csereal.core.news.database
 
-import com.wafflestudio.csereal.common.config.BaseTimeEntity
+import com.wafflestudio.csereal.common.entity.BaseTimeEntity
 import jakarta.persistence.Entity
 import jakarta.persistence.EnumType
 import jakarta.persistence.Enumerated
