+++ src/main/kotlin/com/wafflestudio/csereal/core/recruit/database/RecruitEntity.kt
@@ -1,6 +1,6 @@
 package com.wafflestudio.csereal.core.recruit.database
 
-import com.wafflestudio.csereal.common.config.BaseTimeEntity
+import com.wafflestudio.csereal.common.entity.BaseTimeEntity
 import com.wafflestudio.csereal.common.controller.MainImageContentEntityType
 import com.wafflestudio.csereal.core.resource.mainImage.database.MainImageEntity
 import jakarta.persistence.Column
