+++ src/main/kotlin/com/wafflestudio/csereal/core/research/database/ResearchEntity.kt
@@ -1,6 +1,6 @@
 package com.wafflestudio.csereal.core.research.database
 
-import com.wafflestudio.csereal.common.config.BaseTimeEntity
+import com.wafflestudio.csereal.common.entity.BaseTimeEntity
 import com.wafflestudio.csereal.common.controller.MainImageContentEntityType
 import com.wafflestudio.csereal.common.enums.LanguageType
 import com.wafflestudio.csereal.core.research.dto.ResearchDto
