+++ src/main/kotlin/com/wafflestudio/csereal/core/resource/mainImage/database/MainImageEntity.kt
@@ -1,6 +1,6 @@
 package com.wafflestudio.csereal.core.resource.mainImage.database
 
-import com.wafflestudio.csereal.common.config.BaseTimeEntity
+import com.wafflestudio.csereal.common.entity.BaseTimeEntity
 import jakarta.persistence.*
 
 @Entity(name = "mainImage")
